--- a/.jshintignore
+++ b/.jshintignore
@@ -1,6 +1,0 @@
-node_modules/**
-coverage/**
-**.md
-**.log
-.idea
-benchmark