--- a/.npmignore
+++ b/.npmignore
@@ -1,10 +1,12 @@
-# Created by .ignore support plugin (hsz.mobi)
-### Example user template template
-### Example user template
-
 # IntelliJ project files
 .idea
 *.iml
 out
 gen
+
+# Unrelevant files and folders
 benchmark
+coverage
+test
+.travis.yml
+.gitignore