--- a/index.js
+++ b/index.js
@@ -1,3 +1,4 @@
 'use strict';
 
 module.exports = require('./lib/parser');
+module.exports.ReplyError = require('./lib/replyError');