--- a/index.js
+++ b/index.js
@@ -1,4 +1,4 @@
-'use strict';
+'use strict'
 
-module.exports = require('./lib/parser');
-module.exports.ReplyError = require('./lib/replyError');
+module.exports = require('./lib/parser')
+module.exports.ReplyError = require('./lib/replyError')