--- a/lib/hiredis.js
+++ b/lib/hiredis.js
@@ -25,9 +25,6 @@ function parseData (parser) {
  * @constructor
  */
 function HiredisReplyParser (options) {
-  if (!(this instanceof HiredisReplyParser)) {
-    return new HiredisReplyParser(options)
-  }
   this.returnError = options.returnError
   this.returnFatalError = options.returnFatalError || options.returnError
   this.returnReply = options.returnReply