--- a/lib/javascript.js
+++ b/lib/javascript.js
@@ -1,172 +1,262 @@
 'use strict';
+/*jshint latedef: nofunc */
 
-function JavascriptReplyParser(options) {
-    this.name = 'javascript';
-    this.buffer = new Buffer(0);
-    this.offset = 0;
-    this.bigStrSize = 0;
-    this.chunksSize = 0;
-    this.buffers = [];
-    this.type = 0;
-    this.protocolError = false;
-    this.offsetCache = 0;
-    // If returnBuffers is active, all return values are returned as buffers besides numbers and errors
-    if (options.return_buffers) {
-        this.handleReply = function (start, end) {
-            return this.buffer.slice(start, end);
-        };
-    } else {
-        this.handleReply = function (start, end) {
-            return this.buffer.toString('utf-8', start, end);
-        };
-    }
-    // If stringNumbers is activated the parser always returns numbers as string
-    // This is important for big numbers (number > Math.pow(2, 53)) as js numbers are 64bit floating point numbers with reduced precision
-    if (options.string_numbers) {
-        this.handleNumbers = function (start, end) {
-            return this.buffer.toString('ascii', start, end);
-        };
-    } else {
-        this.handleNumbers = function (start, end) {
-            return +this.buffer.toString('ascii', start, end);
-        };
-    }
+/**
+ * Used for lengths only, faster perf on arrays / bulks
+ * @param parser
+ * @returns {*}
+ */
+function parseSimpleString(parser) {
+	var offset = parser.offset;
+	var length = parser.buffer.length;
+	var string = '';
+
+	while (offset < length) {
+		var c1 = parser.buffer[offset++];
+		if (c1 === 13) {
+			var c2 = parser.buffer[offset++];
+			if (c2 === 10) {
+				parser.offset = offset;
+				return string;
+			}
+			string += String.fromCharCode(c1) + String.fromCharCode(c2);
+			continue;
+		}
+		string += String.fromCharCode(c1);
+	}
+	return undefined;
 }
 
-JavascriptReplyParser.prototype.parseResult = function (type) {
-    var start = 0,
-        end = 0,
-        packetHeader = 0,
-        reply;
+/**
+ * Returns a string or buffer of the provided offset start and
+ * end ranges. Checks `optionReturnBuffers`.
+ * @param parser
+ * @param start
+ * @param end
+ * @returns {*}
+ */
+function convertBufferRange(parser, start, end) {
+	// If returnBuffers is active, all return values are returned as buffers besides numbers and errors
+	if (parser.optionReturnBuffers === true) {
+		return parser.buffer.slice(start, end);
+	}
 
-    if (type === 36) { // $
-        packetHeader = this.parseHeader();
-        // Packets with a size of -1 are considered null
-        if (packetHeader === -1) {
-            return null;
-        }
-        end = this.offset + packetHeader;
-        start = this.offset;
-        if (end + 2 > this.buffer.length) {
-            this.buffers.push(this.offsetCache === 0 ? this.buffer : this.buffer.slice(this.offsetCache));
-            this.chunksSize = this.buffers[0].length;
-            // Include the packetHeader delimiter
-            this.bigStrSize = packetHeader + 2;
-            throw new Error('Wait for more data.');
-        }
-        // Set the offset to after the delimiter
-        this.offset = end + 2;
-        return this.handleReply(start, end);
-    } else if (type === 58) { // :
-        // Up to the delimiter
-        end = this.packetEndOffset();
-        start = this.offset;
-        // Include the delimiter
-        this.offset = end + 2;
-        // Return the coerced numeric value
-        return this.handleNumbers(start, end);
-    } else if (type === 43) { // +
-        end = this.packetEndOffset();
-        start = this.offset;
-        this.offset = end + 2;
-        return this.handleReply(start, end);
-    } else if (type === 42) { // *
-        packetHeader = this.parseHeader();
-        if (packetHeader === -1) {
-            return null;
-        }
-        reply = [];
-        for (var i = 0; i < packetHeader; i++) {
-            if (this.offset >= this.buffer.length) {
-                throw new Error('Wait for more data.');
-            }
-            reply.push(this.parseResult(this.buffer[this.offset++]));
-        }
-        return reply;
-    } else if (type === 45) { // -
-        end = this.packetEndOffset();
-        start = this.offset;
-        this.offset = end + 2;
-        return new Error(this.buffer.toString('utf-8', start, end));
-    }
+	return parser.buffer.toString('utf-8', start, end);
+}
+
+/**
+ * Parse a '+' redis simple string response but forward the offsets
+ * onto convertBufferRange to generate a string.
+ * @param parser
+ * @returns {*}
+ */
+function parseSimpleStringViaOffset(parser) {
+	var start = parser.offset;
+	var offset = parser.offset;
+	var length = parser.buffer.length;
+
+	while (offset < length) {
+		var c1 = parser.buffer[offset++];
+		if (c1 === 13) { // \r
+			var c2 = parser.buffer[offset++];
+			if (c2 === 10) { // \n
+				parser.offset = offset;
+				return convertBufferRange(parser, start, offset - 2);
+			}
+		}
+	}
+	return undefined;
+}
+
+/**
+ * Returns the string length via parseSimpleString
+ * @param parser
+ * @returns {*}
+ */
+function parseLength(parser) {
+	var string = parseSimpleString(parser);
+	if (string !== undefined) {
+		var length = +string;
+		if (length === -1) {
+			return null;
+		}
+		return length;
+	}
+}
+
+/**
+ * Parse a ':' redis integer response
+ * @param parser
+ * @returns {*}
+ */
+function parseInteger(parser) {
+	var string = parseSimpleStringViaOffset(parser);
+	if (string !== undefined) {
+		// If stringNumbers is activated the parser always returns numbers as string
+		// This is important for big numbers (number > Math.pow(2, 53)) as js numbers
+		// are 64bit floating point numbers with reduced precision
+		if (parser.optionStringNumbers === false) {
+			return +string;
+		}
+		return string;
+	}
+}
+
+/**
+ * Parse a '$' redis bulk string response
+ * @param parser
+ * @returns {null}
+ */
+function parseBulkString(parser) {
+	var length = parseLength(parser);
+	/* jshint eqnull: true */
+	if (length == null) {
+		return length;
+	}
+	var offsetEnd = parser.offset + length;
+	if ((offsetEnd + 2) > parser.buffer.length) {
+		return;
+	}
+
+	var offsetBegin = parser.offset;
+	parser.offset = offsetEnd + 2;
+
+	return convertBufferRange(parser, offsetBegin, offsetEnd);
+}
+
+/**
+ * Parse a '-' redis error response
+ * @param parser
+ * @returns {Error}
+ */
+function parseError(parser) {
+	var string = parseSimpleStringViaOffset(parser);
+	if (string !== undefined) {
+		return new Error(string);
+	}
+}
+
+/**
+ * Parsing error handler, resets parser buffer
+ * @param parser
+ * @param error
+ */
+function handleError(parser, error) {
+	parser.buffer = null;
+	parser.returnFatalError(error);
+}
+
+/**
+ * Parse a
+ * @param parser
+ * @returns {*}
+ */
+function parseArray(parser) {
+	var length = parseLength(parser);
+	/* jshint eqnull: true */
+	if (length == null) { // will break if using ===
+		return length;
+	}
+
+	var responses = new Array(length);
+	var bufferLength = parser.buffer.length;
+	for (var i = 0; i < length; i++) {
+		if (parser.offset >= bufferLength) {
+			return;
+		}
+		var response = parseType(parser, parser.buffer[parser.offset++]);
+		if (response === undefined) {
+			return;
+		}
+		responses[i] = response;
+	}
+
+	return responses;
+}
+
+/**
+ * Called the appropriate parser for the specified type.
+ * @param parser
+ * @param type
+ * @returns {*}
+ */
+function parseType(parser, type) {
+	switch (type) {
+		case 36: // $
+			return parseBulkString(parser);
+		case 58: // :
+			return parseInteger(parser);
+		case 43: // +
+			return parser.optionReturnBuffers ? parseSimpleStringViaOffset(parser) : parseSimpleString(parser);
+		case 42: // *
+			return parseArray(parser);
+		case 45: // -
+			return parseError(parser);
+		default:
+			return handleError(parser, new Error('Protocol error, got ' + JSON.stringify(String.fromCharCode(type)) + ' as reply type byte'));
+	}
+}
+
+/**
+ * Quick buffer appending via buffer copy.
+ * @param parser
+ * @param buffer
+ */
+function appendToBuffer(parser, buffer) {
+	var oldLength = parser.buffer.length;
+	var remainingLength = oldLength - parser.offset;
+	var newLength = remainingLength + buffer.length;
+	var newBuffer = new Buffer(newLength);
+	parser.buffer.copy(newBuffer, 0, parser.offset, oldLength);
+	buffer.copy(newBuffer, remainingLength, 0, buffer.length);
+	parser.buffer = newBuffer;
+	parser.offset = 0;
+}
+
+/**
+ * Javascript Redis Parser
+ * @param options
+ * @constructor
+ */
+function JavascriptRedisParser(options) {
+	this.optionReturnBuffers = !!options.return_buffers;
+	this.optionStringNumbers = !!options.string_numbers;
+	this.name = 'javascript';
+	this.offset = 0;
+	this.buffer = null;
+}
+
+/**
+ * Parse the redis buffer
+ * @param buffer
+ */
+JavascriptRedisParser.prototype.execute = function (buffer) {
+	if (this.buffer === null) {
+		this.buffer = buffer;
+		this.offset = 0;
+	} else {
+		appendToBuffer(this, buffer);
+	}
+
+	var length = this.buffer.length;
+
+	while (this.offset < length) {
+		var offset = this.offset;
+		var type = this.buffer[this.offset++];
+		var response = parseType(this, type);
+		if (response === undefined) {
+			this.offset = offset;
+			return;
+		}
+
+		if (type === 45) {
+			this.returnError(response); // Errors -
+		} else {
+			this.returnReply(response); // Strings + // Integers : // Bulk strings $ // Arrays *
+		}
+	}
+
+	this.buffer = null;
 };
 
-JavascriptReplyParser.prototype.execute = function (buffer) {
-    if (this.chunksSize !== 0) {
-        if (this.bigStrSize > this.chunksSize + buffer.length) {
-            this.buffers.push(buffer);
-            this.chunksSize += buffer.length;
-            return;
-        }
-        this.buffers.push(buffer);
-        this.buffer = Buffer.concat(this.buffers, this.chunksSize + buffer.length);
-        this.buffers = [];
-        this.bigStrSize = 0;
-        this.chunksSize = 0;
-    } else if (this.offset >= this.buffer.length) {
-        this.buffer = buffer;
-    } else {
-        this.buffer = Buffer.concat([this.buffer.slice(this.offset), buffer]);
-    }
-    this.offset = 0;
-    this.run();
-};
-
-JavascriptReplyParser.prototype.tryParsing = function () {
-    try {
-        return this.parseResult(this.type);
-    } catch (err) {
-        // Catch the error (not enough data), rewind if it's an array,
-        // and wait for the next packet to appear
-        this.offset = this.offsetCache;
-        // Indicate that there's no protocol error by resetting the type too
-        this.type = undefined;
-    }
-};
-
-JavascriptReplyParser.prototype.run = function () {
-    // Set a rewind point. If a failure occurs, wait for the next execute()/append() and try again
-    this.offsetCache = this.offset;
-    this.type = this.buffer[this.offset++];
-    var reply = this.tryParsing();
-
-    while (reply !== undefined) {
-        if (this.type === 45) { // Errors -
-            this.returnError(reply);
-        } else {
-            this.returnReply(reply); // Strings + // Integers : // Bulk strings $ // Arrays *
-        }
-        this.offsetCache = this.offset;
-        this.type = this.buffer[this.offset++];
-        reply = this.tryParsing();
-    }
-    if (this.type !== undefined) {
-        // Reset the buffer so the parser can handle following commands properly
-        this.buffer = new Buffer(0);
-        this.returnFatalError(new Error('Protocol error, got ' + JSON.stringify(String.fromCharCode(this.type)) + ' as reply type byte'));
-    }
-};
-
-JavascriptReplyParser.prototype.parseHeader = function () {
-    var end   = this.packetEndOffset(),
-        value = this.buffer.toString('ascii', this.offset, end) | 0;
-
-    this.offset = end + 2;
-    return value;
-};
-
-JavascriptReplyParser.prototype.packetEndOffset = function () {
-    var offset = this.offset,
-        len = this.buffer.length - 1;
-
-    while (this.buffer[offset] !== 0x0d && this.buffer[offset + 1] !== 0x0a) {
-        offset++;
-
-        if (offset >= len) {
-            throw new Error('Did not see LF after NL reading multi bulk count (' + offset + ' => ' + this.buffer.length + ', ' + this.offset + ')');
-        }
-    }
-    return offset;
-};
-
-module.exports = JavascriptReplyParser;
+module.exports = JavascriptRedisParser;