--- a/lib/parser.js
+++ b/lib/parser.js
@@ -145,7 +145,7 @@ function handleError (parser, error) {
 }
 
 /**
- * Parse a
+ * Parse a '*' redis array response
  * @param parser
  * @returns {*}
  */
@@ -234,6 +234,7 @@ function JavascriptRedisParser (options) {
 /**
  * Parse the redis buffer
  * @param buffer
+ * @returns {undefined}
  */
 JavascriptRedisParser.prototype.execute = function (buffer) {
   if (this.buffer === null) {