--- a/lib/parser.js
+++ b/lib/parser.js
@@ -214,6 +214,7 @@ function JavascriptRedisParser (options) {
   if (options.name === 'hiredis') {
     try {
       var Hiredis = require('../test/hiredis')
+      console.error(new TypeError('Using the hiredis parser is discouraged. Please remove the name option.').stack)
       return new Hiredis(options)
     } catch (e) {}
   }