--- a/lib/parser.js
+++ b/lib/parser.js
@@ -210,6 +210,7 @@ function JavascriptRedisParser (options) {
     throw new TypeError('Please provide all return functions while initiating the parser')
   }
   if (options.name === 'hiredis') {
+    /* istanbul ignore next: hiredis is only supported for legacy usage */
     try {
       var Hiredis = require('../test/hiredis')
       console.error(new TypeError('Using the hiredis parser is discouraged. Please remove the name option.').stack.replace('Error', 'Warning'))