--- a/lib/parser.js
+++ b/lib/parser.js
@@ -8,18 +8,25 @@ var ReplyError = require('./replyError')
  * @param parser
  * @returns {*}
  */
-function parseSimpleString (parser) {
+function parseSimpleNumbers (parser) {
   var offset = parser.offset
   var length = parser.buffer.length
   var string = ''
+  var c1 = parser.buffer[offset++]
+
+  if (c1 === 45) {
+    string += '-'
+  } else {
+    string += c1 - 48
+  }
 
   while (offset < length) {
-    var c1 = parser.buffer[offset++]
+    c1 = parser.buffer[offset++]
     if (c1 === 13 && parser.buffer[offset] === 10) { // \r\n
       parser.offset = offset + 1
       return string
     }
-    string += String.fromCharCode(c1)
+    string += c1 - 48
   }
 }
 
@@ -61,12 +68,12 @@ function parseSimpleStringViaOffset (parser) {
 }
 
 /**
- * Returns the string length via parseSimpleString
+ * Returns the string length via parseSimpleNumbers
  * @param parser
  * @returns {*}
  */
 function parseLength (parser) {
-  var string = parseSimpleString(parser)
+  var string = parseSimpleNumbers(parser)
   if (string !== undefined) {
     return +string
   }
@@ -78,7 +85,7 @@ function parseLength (parser) {
  * @returns {*}
  */
 function parseInteger (parser) {
-  var string = parseSimpleString(parser)
+  var string = parseSimpleNumbers(parser)
   if (string !== undefined) {
     // If stringNumbers is activated the parser always returns numbers as string
     // This is important for big numbers (number > Math.pow(2, 53)) as js numbers