--- a/lib/parser.js
+++ b/lib/parser.js
@@ -132,6 +132,9 @@ function parseBulkString (parser) {
 function parseError (parser) {
   var string = parseSimpleStringViaOffset(parser)
   if (string !== undefined) {
+    if (parser.optionReturnBuffers === true) {
+      string = string.toString()
+    }
     return new ReplyError(string)
   }
 }