--- a/lib/parser.js
+++ b/lib/parser.js
@@ -82,10 +82,9 @@ function parseSimpleStringViaOffset (parser) {
   var length = parser.buffer.length
 
   while (offset < length) {
-    var c1 = parser.buffer[offset++]
-    if (c1 === 13 && parser.buffer[offset] === 10) { // \r\n
-      parser.offset = offset + 1
-      return convertBufferRange(parser, start, offset - 1)
+    if (parser.buffer[offset++] === 10) { // \r\n
+      parser.offset = offset
+      return convertBufferRange(parser, start, offset - 2)
     }
   }
 }