--- a/lib/parser.js
+++ b/lib/parser.js
@@ -135,7 +135,7 @@ function parseBulkString (parser) {
   if (offsetEnd + 2 > parser.buffer.length) {
     parser.bufferCache.push(parser.buffer)
     parser.totalChunkSize = parser.buffer.length
-    parser.bigStrSize = length + 2
+    parser.bigStrSize = offsetEnd + 2
     return
   }