This is a complete rewrite of the JS parser done by @Salakar with additional work by me.

As the hiredis parser is now likely slower in every case, I went ahead and deprecated it and removed any hints about it in the documentation. It's still usable by using the undocumented name option, but it'll print a warning in that case.

All errors returned by the parser are from now on "ReplyErrors" with a limited stack, as the stack is not meaningful at all and requires computation time.

```
OLD CODE: multiple chunks in a bulk string x 57,774 ops/sec ±1.07% (91 runs sampled)
HIREDIS: multiple chunks in a bulk string x 899,793 ops/sec ±0.75% (90 runs sampled)
NEW CODE: multiple chunks in a bulk string x 1,618,283 ops/sec ±0.87% (89 runs sampled)
NEW BUF: multiple chunks in a bulk string x 850,057 ops/sec ±0.86% (93 runs sampled)

OLD CODE: multiple chunks in a string x 125,927 ops/sec ±0.91% (93 runs sampled)
HIREDIS: multiple chunks in a string x 1,730,827 ops/sec ±0.46% (95 runs sampled)
NEW CODE: multiple chunks in a string x 1,581,035 ops/sec ±0.59% (91 runs sampled)
NEW BUF: multiple chunks in a string x 1,366,036 ops/sec ±0.61% (93 runs sampled)

OLD CODE: 4mb bulk string x 248 ops/sec ±1.46% (79 runs sampled)
HIREDIS: 4mb bulk string x 362 ops/sec ±1.70% (75 runs sampled)
NEW CODE: 4mb bulk string x 897 ops/sec ±1.52% (77 runs sampled)
NEW BUF: 4mb bulk string x 1,586 ops/sec ±1.77% (90 runs sampled)

OLD CODE: + simple string x 2,546,973 ops/sec ±0.67% (90 runs sampled)
HIREDIS: + simple string x 2,312,779 ops/sec ±0.56% (94 runs sampled)
NEW CODE: + simple string x 4,537,204 ops/sec ±0.83% (91 runs sampled)
NEW BUF: + simple string x 3,016,757 ops/sec ±0.73% (92 runs sampled)

OLD CODE: + integer x 2,728,872 ops/sec ±0.64% (92 runs sampled)
HIREDIS: + integer x 2,191,210 ops/sec ±0.51% (93 runs sampled)
NEW CODE: + integer x 15,273,412 ops/sec ±0.90% (89 runs sampled)
NEW STR: + integer x 6,837,573 ops/sec ±1.06% (87 runs sampled)

OLD CODE: + big integer x 1,786,652 ops/sec ±0.60% (94 runs sampled)
HIREDIS: + big integer x 2,041,817 ops/sec ±0.45% (95 runs sampled)
NEW CODE: + big integer x 7,073,548 ops/sec ±0.99% (89 runs sampled)
NEW STR: + big integer x 3,094,097 ops/sec ±0.99% (88 runs sampled)

OLD CODE: * array x 1,294,499 ops/sec ±0.63% (94 runs sampled)
HIREDIS: * array x 941,297 ops/sec ±0.52% (96 runs sampled)
NEW CODE: * array x 4,037,489 ops/sec ±0.81% (86 runs sampled)
NEW BUF: * array x 2,635,257 ops/sec ±0.68% (90 runs sampled)

OLD CODE: * bigArray x 35,453 ops/sec ±0.74% (96 runs sampled)
HIREDIS: * bigArray x 38,992 ops/sec ±0.68% (91 runs sampled)
NEW CODE: * bigArray x 67,108 ops/sec ±0.65% (95 runs sampled)
NEW BUF: * bigArray x 39,251 ops/sec ±0.81% (92 runs sampled)

OLD CODE: * error x 170,299 ops/sec ±1.11% (85 runs sampled)
HIREDIS: * error x 88,664 ops/sec ±0.83% (92 runs sampled)
NEW CODE: * error x 179,716 ops/sec ±0.72% (92 runs sampled)
```

The benchmark was run on a Lenovo T450s with an i7 on NodeJS 6.1.

Supersedes and closes #2 

__Edit__: I updated the benchmark numbers and the the comment about removing hiredis, as it's currently only deprecated to keep backwards compatibility.