@luin 2016-05-09
I ran the new parser with ioredis, and the benchmark result is impressive. Awesome work!

However, when parsing a big array, new parser seems to be slower than the old one:

```javascript
var bigArraySize = 10;
var bigArray = '*' + bigArraySize + '\r\n';
for (var i = 0; i < bigArraySize; i++) {
  bigArray += '$';
  var size = (Math.random() * 10 | 0) + 1;
  bigArray += size + '\r\n' + lorem.slice(0, size) + '\r\n';
}

var bigArrayList = bigArray.split('\r\n').map(function (str) {
  return new Buffer(str);
});

// BIG ARRAYS

suite.add('\nOLD CODE: * bigArray', function () {
  for (var i = 0; i < bigArrayList.length; i++) {
    parserOld.execute(bigArrayList[i]);
  }
})

suite.add('HIREDIS: * bigArray', function () {
  for (var i = 0; i < bigArrayList.length; i++) {
    parserHiRedis.execute(bigArrayList[i]);
  }
})

suite.add('NEW CODE: * bigArray', function () {
  for (var i = 0; i < bigArrayList.length; i++) {
    parser.execute(bigArrayList[i]);
  }
})
```

The result:

```
OLD CODE: * bigArray x 251 ops/sec ±30.59% (15 runs sampled)
HIREDIS: * bigArray x 459 ops/sec ±77.80% (5 runs sampled)
NEW CODE: * bigArray x 68.65 ops/sec ±83.74% (5 runs sampled)
```