@BridgeAR 2016-05-09
The array example is not good, as all the `\r\n` are removed and it also mixes the array itself with lots of small chunks that have to be collected first.