@BridgeAR 2016-05-17
I updated the benchmark numbers and they look very promising to me :smiley:

I'm still wondering why `NEW CODE: + simple string` has decreased for me from 5 million to 4.5 million. That is the only drop with my further improvements (mainly about multiple chunks).

Currently I think it's difficult to get much better results than the ones we reached. @Salakar @luin do you mind doing another code review? I can move the new commits to another branch if you wish.