@Salakar 2016-05-17
As per our call, this is all good, so I say LGTM and released, unless @luin has any issues?

Further perf tweaks can be done in v2 minor versions if needed.