@luin 2016-05-17
I just ran the benchmark and it looks good to me. 👍