@Salakar 2016-05-17
@luin what's the perf increases like on ioredis?