@BridgeAR 2016-05-22
No, there are regressions left while concatenating bulk strings that I addressed locally but I want to rewrite it a bit more but currently I'm occupied by [hack 4 humanity](https://www.worldhumanitariansummit.org/summit/specialevents/hack4humanity).