@luin 2016-05-26
@Salakar I'm going to do more tests this weekend. If the new parser is fast enough that we don't need to introduce `dropBufferSupport`, I'll definitely wait for it. 😆